requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
transformers==4.37.2
gTTS==2.5.1
python-multipart==0.0.9
httpx
python-dotenv